IO now compiles. Back to debugging initialization.

+++ make.sh
@@ -2,12 +2,18 @@
 # Used to decide flags and call cmake
 # TODO autodetection?  Machinefiles at least?
 
-rm -rf build
-mkdir build
+if [ "$1" == "clean" ]; then
+  rm -rf build
+  mkdir build
+fi
+
 cd build
 
+if [ "$1" == "clean" ]; then
 cmake3 ..\
     -DCMAKE_CXX_COMPILER=$PWD/../external/kokkos/bin/nvcc_wrapper \
+    -DCMAKE_BUILD_TYPE=Debug \
+    -DCMAKE_CXX_FLAGS_DEBUG="-DDEBUG -Wall" \
     -DUSE_MPI=OFF \
     -DKokkos_ENABLE_OPENMP=ON \
     -DKokkos_ENABLE_CUDA=ON \
@@ -20,6 +26,7 @@ cmake3 ..\
     -DKokkos_ARCH_MAXWELL52=OFF \
     -DKokkos_ARCH_VOLTA70=ON \
     -DKokkos_ENABLE_CUDA_LAMBDA=ON
+fi
 
 make -j
 cp kharm/kharm.* ..
